+++ docs/guides/changelog.md
@@ -12,11 +12,12 @@ All notable changes to this project will be documented in this file. All dates a
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [v0.0.next] - xx-xx-2026
+## [next-release] - xx-xx-2026
 
 ### Added
 - Added data sources for Authentication Methods.
 - Added data sources for Certificate Trust Lists.
+- Added data sources for Enforcement Policies.
 
 ## [v0.0.7] - 26-02-2026